BackDoggieHommie/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
psycopg2-binary==2.9.3
sqlparse==0.4.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Pillow==9.1.0
django-cors-headers==3.11.0
six==1.16.0
pyrebase===3.0.27